Optimise circle detection accuracy
Scaling the image reduces cpu time when using canny detection. But
rescaling back may introduce an offset.

This PR realigns the centroid as well as the diameter of the perimiter
by using the rescaled perimiter points as an initial guess:
- Search along the line between the center point and 8 points (initial
guess) towards and outwards the center starting at the inital guess
- Find a threshold between 2 consecutive points (black/white)
- Fit a circle between those adjusted points